move some pane structs into `data` crate
