+++ .gitignore
@@ -2,4 +2,3 @@
 /target
 /.vscode
 *.log
-/data
