+++ data/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "data"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+serde_json.workspace = true
+serde.workspace = true
+exchanges = { version = "0.1.0", path = "../exchanges" }
