+++ data/src/charts/heatmap.rs
@@ -1,0 +1,20 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Copy, Clone, PartialEq, Deserialize, Serialize)]
+pub struct Config {
+    pub trade_size_filter: f32,
+    pub order_size_filter: f32,
+    pub dynamic_sized_trades: bool,
+    pub trade_size_scale: i32,
+}
+
+impl Default for Config {
+    fn default() -> Self {
+        Config {
+            trade_size_filter: 0.0,
+            order_size_filter: 0.0,
+            dynamic_sized_trades: true,
+            trade_size_scale: 100,
+        }
+    }
+}
