+++ data/src/pane.rs
@@ -1,0 +1,55 @@
+use exchanges::{TickMultiplier, TickerInfo, adapter::StreamType};
+use serde::{Deserialize, Serialize};
+
+use crate::charts::{
+    ChartBasis, ChartLayout, VisualConfig,
+    indicators::{CandlestickIndicator, FootprintIndicator, HeatmapIndicator},
+};
+
+#[derive(Debug, Clone, Copy, Deserialize, Serialize)]
+pub enum Axis {
+    Horizontal,
+    Vertical,
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub enum SerializablePane {
+    Split {
+        axis: Axis,
+        ratio: f32,
+        a: Box<SerializablePane>,
+        b: Box<SerializablePane>,
+    },
+    Starter,
+    HeatmapChart {
+        layout: ChartLayout,
+        stream_type: Vec<StreamType>,
+        settings: PaneSettings,
+        indicators: Vec<HeatmapIndicator>,
+    },
+    FootprintChart {
+        layout: ChartLayout,
+        stream_type: Vec<StreamType>,
+        settings: PaneSettings,
+        indicators: Vec<FootprintIndicator>,
+    },
+    CandlestickChart {
+        layout: ChartLayout,
+        stream_type: Vec<StreamType>,
+        settings: PaneSettings,
+        indicators: Vec<CandlestickIndicator>,
+    },
+    TimeAndSales {
+        stream_type: Vec<StreamType>,
+        settings: PaneSettings,
+    },
+}
+
+#[derive(Debug, Clone, Copy, Deserialize, Serialize, Default)]
+#[serde(default)]
+pub struct PaneSettings {
+    pub ticker_info: Option<TickerInfo>,
+    pub tick_multiply: Option<TickMultiplier>,
+    pub visual_config: Option<VisualConfig>,
+    pub selected_basis: Option<ChartBasis>,
+}
