+++ exchanges/Cargo.toml
@@ -9,6 +9,7 @@ chrono.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 log.workspace = true
+rust_decimal.workspace = true
 thiserror.workspace = true
 ordered-float.workspace = true
 regex.workspace = true
