+++ src/charts.rs
@@ -1,3 +1,4 @@
+use data::charts::indicators::Indicator;
 use iced::widget::canvas::{self, Cache, Canvas, Event, Frame};
 use iced::widget::{center, mouse_area};
 use iced::{
@@ -10,9 +11,7 @@ use iced::{
         tooltip::Position as TooltipPosition,
     },
 };
-use indicators::Indicator;
 use scales::{AxisLabelsX, AxisLabelsY, PriceInfoLabel};
-use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
 use crate::{
@@ -23,7 +22,7 @@ use crate::{
     widget::hsplit::HSplit,
     widget::tooltip,
 };
-use data::charts::ChartLayout;
+use data::charts::{ChartBasis, ChartLayout};
 use exchanges::{TickerInfo, Timeframe};
 
 pub mod candlestick;
@@ -504,49 +503,6 @@ impl Caches {
         self.x_labels.clear();
         self.y_labels.clear();
         self.crosshair.clear();
-    }
-}
-
-/// Defines how chart data is aggregated and displayed along the x-axis.
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-pub enum ChartBasis {
-    /// Time-based aggregation where each datapoint represents a fixed time interval.
-    ///
-    /// The u64 value represents milliseconds. Common values include:
-    /// - `60_000` (1 minute)
-    /// - `300_000` (5 minutes)
-    /// - `3_600_000` (1 hour)
-    Time(u64),
-
-    /// Trade-based aggregation where each datapoint represents a fixed number of trades.
-    ///
-    /// The u64 value represents the number of trades per aggregation unit.
-    /// Common values include 100, 500, or 1000 trades per bar/candle.
-    Tick(u64),
-}
-
-impl ChartBasis {
-    pub fn is_time(&self) -> bool {
-        matches!(self, ChartBasis::Time(_))
-    }
-}
-
-impl std::fmt::Display for ChartBasis {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ChartBasis::Time(millis) => match *millis {
-                60_000 => write!(f, "1m"),
-                180_000 => write!(f, "3m"),
-                300_000 => write!(f, "5m"),
-                900_000 => write!(f, "15m"),
-                1_800_000 => write!(f, "30m"),
-                3_600_000 => write!(f, "1h"),
-                7_200_000 => write!(f, "2h"),
-                14_400_000 => write!(f, "4h"),
-                _ => write!(f, "{millis}ms"),
-            },
-            ChartBasis::Tick(count) => write!(f, "{count}T"),
-        }
     }
 }
